Fixed a few warnings.

+++ src/las.h
@@ -151,7 +151,7 @@ void print_tree(node *n);
 struct used_vars;
 void fix_type_field(node *n);
 struct timer_oflo;
-int eval_low(node *n);
+ptrdiff_t eval_low(node *n);
 int dooptcode(struct pike_string *name,
 	      node *n,
 	      struct pike_string *type,
